Moved from JST to Handlebars for template compilation
Note that this commit does not include nesting views, and breaks
compilation of nesting views entirely.

View templates are now processed separately as handlebars files, and
compiled using the handlebars pre-compiler then prepended to
`application.js`. The pre-compile stores the templates on
`window.Handlebars.templates` as functions that, like JST, take a
context object containing key/values to interpolate in the template.
This means the templates can be compiled in Views as before
(`@template(context)`).

The generation and compile UX are unaffected by these changes.